backend/requirements.txt
# DPDP GUI Compliance Scanner - Backend Dependencies

# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg>=0.28.0
# psycopg2-binary==2.9.9
alembic==1.13.1

# Task Queue
celery==5.3.6
redis==5.0.1

# Web Scanning
playwright==1.41.2
beautifulsoup4==4.12.3
lxml>=5.0.0
httpx==0.26.0

# Windows Scanning
pywinauto==0.6.8
opencv-python==4.9.0.80
pytesseract==0.3.10
Pillow>=10.0.0

# NLP
spacy==3.7.4
langdetect==1.0.9
textstat==0.7.3

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Storage
minio==7.2.3

# Reporting
reportlab==4.1.0
openpyxl==3.1.2
jinja2==3.1.3

# WebSocket
websockets==12.0

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
pyyaml==6.0.1
tenacity==8.2.3

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
